# server/Swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "To do api"
host: "localhost:8080"
tags:
  - name: "Todolist"
    description: "Operations with to-do lists"
  - name: "Tasks"
    description: "Operations with tasks"
  - name: "User"
    description: "Operations about user"
schemes:
  - "http"
  - "https"
paths:
  /todolists:
    post:
      tags:
        - "Todolist"
      summary: "Add new todo"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/Todolist"
      responses:
        "201":
          description: "Created"
          schema:
            $ref: "#/definitions/TodolistResponse"
        "421":
          description: "No title or status"
          schema:
            $ref: "#/definitions/Error"
      security:
        - Bearer: []
    get:
      tags:
        - "Todolist"
      summary: "Get todos"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/TodolistResponse"
        "204":
          description: "No todo"
          schema:
            $ref: "#/definitions/Error"
      security:
        - Bearer: []
  /todolists/{id}:
    delete:
      tags:
        - "Todolist"
      summary: "Delete todo"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "Success"
      security:
        - Bearer: []
    put:
      tags:
        - "Todolist"
      summary: "Update  todo"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/Todolist"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/TodolistResponse"
        "421":
          description: "No title"
          schema:
            $ref: "#/definitions/Error"
      security:
        - Bearer: []
  /todolists/{id}/tasks:
    post:
      tags:
        - "Tasks"
      summary: "Add task"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/Task"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/TaskResponse"
        "421":
          description: "No title"
          schema:
            $ref: "#/definitions/Error"
      security:
        - Bearer: []
    get:
      tags:
        - "Tasks"
      summary: "Get tasks"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/TaskResponse"
        "204":
          description: "No todo"
          schema:
            $ref: "#/definitions/Error"
      security:
        - Bearer: []
  /todolists/{id}/tasks/{task_id}:
    delete:
      tags:
        - "Tasks"
      summary: "Delete task"
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
          format: "int64"
        - name: "task_id"
          in: "path"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "Success"
      security:
        - Bearer: []
    put:
      tags:
        - "Tasks"
      summary: "Update task"
      description: "Status key-value: 0 - New, 1 - Active, 2 - Done "
      parameters:
        - name: "id"
          in: "path"
          required: true
          type: "integer"
          format: "int64"
        - name: "task_id"
          in: "path"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/UpdatedTask"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/TaskResponse"
        "421":
          description: "No title"
          schema:
            $ref: "#/definitions/Error"
      security:
        - Bearer: []
  /register:
    post:
      tags:
        - "User"
      summary: "Create user"
      description: "A password should be alphanumeric.First letter of the password should be capital.Password must contain a special character (@, $, !, &, etc).Password length must be greater than 8 characters"
      parameters:
        - in: "body"
          name: "body"
          description: "Created user object"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "204":
          description: "Success"
        "403":
          description: "User exist"
          schema:
            $ref: "#/definitions/Error"
        "421":
          description: "Invalid password"
          schema:
            $ref: "#/definitions/Error"
  /login:
    post:
      tags:
        - "User"
      summary: "User login"
      parameters:
        - in: "body"
          name: "body"
          description: "Created user object"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/JwtResponse"
        "401":
          description: "Invalid password or name"
          schema:
            $ref: "#/definitions/Error"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
  User:
    type: "object"
    properties:
      name:
        type: "string"
      password:
        type: "string"
  Todolist:
    type: "object"
    properties:
      title:
        type: "string"
  Task:
    type: "object"
    properties:
      title:
        type: "string"
  UpdatedTask:
    type: "object"
    properties:
      title:
        type: "string"
      status:
        type: "integer"
        format: "int64"
  Error:
    type: "object"
    properties:
      statusCode:
        type: "integer"
        format: "int32"
      err:
        type: "string"
  TodolistResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      user_id:
        $ref: "#/definitions/identificator"
      title:
        type: "string"
  TaskResponse:
    description: "Status key-value: 0 - New, 1 - Active, 2 - Done "
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      user_id:
        $ref: "#/definitions/identificator"
      title:
        type: "string"
      todolist_id:
        type: "integer"
        format: "int32"
      status:
        type: "integer"
        format: "int32"
  JwtResponse:
    type: "object"
    properties:
      token:
        type: "string"
  identificator:
    type: "integer"